+++ R/recycle_data.R
@@ -24,6 +24,7 @@
 #' @param type Column name of column with types
 #' @param default_scenario name of the default scenario.
 #' @param scenario_reported_estimated name of the reported/estimated scenario.
+#' @param scenario_covid_shock name of the scenario with the COVID-19 shock years.
 #' @param scenario_reference_infilling name of the WHO technical programs projections/imputations scenario.
 #' @param include_projection Boolean to include or not projections in recycling
 #' @param recycle_campaigns Boolean to include or not campaigns in recycling
@@ -47,6 +48,7 @@ recycle_data <- function(df,
                          scenario = "scenario",
                          default_scenario = "default",
                          scenario_reported_estimated = "routine",
+                         scenario_covid_shock = "covid_shock",
                          scenario_reference_infilling = "reference_infilling",
                          include_projection = TRUE,
                          recycle_campaigns = TRUE,
@@ -54,7 +56,6 @@ recycle_data <- function(df,
                          trim_years = TRUE) {
   assert_columns(df, iso3, ind, value, year, scenario, type)
   assert_unique_rows(df, ind, iso3, year, scenario = scenario, ind_ids)
-  assert_scenario_in_df(df, c(scenario_reported_estimated, scenario_reference_infilling), scenario)
 
   billion <- rlang::arg_match(billion)
   if (is.null(ind_ids)) {
@@ -79,6 +80,7 @@ recycle_data <- function(df,
       scenario_col = scenario,
       default_scenario = default_scenario,
       scenario_reported_estimated = scenario_reported_estimated,
+      scenario_covid_shock = scenario_covid_shock,
       scenario_reference_infilling = scenario_reference_infilling,
       include_projection = include_projection,
       recycle_campaigns = recycle_campaigns,
@@ -87,19 +89,22 @@ recycle_data <- function(df,
     )
   )
 }
+
 #' Recycle data between scenarios for a single scenario
 #'
 #' `recycle_data_scenario_single ` reuses values present in the specified
-#' scenarios in `default_scenario`, `scenario_reported_estimated`, and
-#' `scenario_reference_infilling` for the specified scenarios.
+#' scenarios in `default_scenario`, `scenario_reported_estimated`,
+#' `scenario_covid_shock` and `scenario_reference_infilling` for the specified
+#' scenarios.
 #'
 #' To do so, it looks at:
 #'
 #' 1. values in `default_scenario` but not in the scenario specified
-#' 2. values in `scenario_reported_estimated` but not in the scenario specified
-#' or `scenario_reference_infilling`
+#' 2. values in `scenario_reported_estimated` or `scenario_covid_shock` but not
+#' in the scenario specified or `default_scenario`.
 #' 3. values in `scenario_reference_infilling` but not in the scenario specified,
-#' `scenario_reported_estimated` or `scenario_reference_infilling`
+#' `scenario_reported_estimated`, `scenario_covid_shock`, or
+#' `scenario_reference_infilling`
 #'
 #' For more information see:
 #'
@@ -112,6 +117,7 @@ recycle_data <- function(df,
 #' @param default_scenario name of the default scenario.
 #' @param scenario_reported_estimated name of the reported/estimated scenario.
 #' @param scenario_reference_infilling name of the WHO technical programs projections/imputations scenario.
+#' @param scenario_covid_shock name of the scenario with the COVID-19 shock years.
 #' @param include_projection Boolean to include or not projections in recycling
 #' @param recycle_campaigns Boolean to include or not campaigns in recycling
 #'
@@ -134,6 +140,7 @@ recycle_data_scenario_single <- function(df,
                                          scenario_col = "scenario",
                                          default_scenario = "default",
                                          scenario_reported_estimated = "routine",
+                                         scenario_covid_shock = "covid_shock",
                                          scenario_reference_infilling = "reference_infilling",
                                          include_projection = TRUE,
                                          recycle_campaigns = TRUE,
@@ -141,11 +148,16 @@ recycle_data_scenario_single <- function(df,
                                          trim_years = FALSE) {
   assert_columns(df, iso3, ind, value, year, scenario_col, type)
   assert_unique_rows(df, ind, iso3, year, scenario = scenario_col, ind_ids)
-  assert_scenario_in_df(df, c(scenario, scenario_reported_estimated, scenario_reference_infilling), scenario_col)
 
   billion <- rlang::arg_match(billion)
   if (is.null(ind_ids)) {
     ind_ids <- billion_ind_codes(billion)
+  }
+
+  if (scenario %in% c(scenario_reported_estimated, scenario_covid_shock, scenario_reference_infilling)) {
+    df_no_recycling <- df %>%
+      dplyr::filter(.data[[scenario_col]] == !!scenario)
+    return(df_no_recycling)
   }
 
   assert_ind_ids(ind_ids, billion)
@@ -155,6 +167,9 @@ recycle_data_scenario_single <- function(df,
 
   reported_estimated_df <- df %>%
     dplyr::filter(.data[[scenario_col]] == !!scenario_reported_estimated)
+
+  covid_shock_df <- df %>%
+    dplyr::filter(.data[[scenario_col]] == !!scenario_covid_shock)
 
   reference_infilling_df <- df %>%
     dplyr::filter(.data[[scenario_col]] == !!scenario_reference_infilling)
@@ -174,7 +189,16 @@ recycle_data_scenario_single <- function(df,
     by = c(iso3, ind, year)
   )
 
-  reference_infilling_not_in_scenario <- dplyr::anti_join(reference_infilling_df, scenario_df,
+  covid_shock_not_in_scenario <- dplyr::anti_join(covid_shock_df, scenario_df,
+    by = c(iso3, ind, year)
+  )
+
+  covid_shock_not_in_default <- dplyr::anti_join(covid_shock_not_in_scenario, default_not_in_scenario,
+    by = c(iso3, ind, year)
+  )
+
+  reference_infilling_not_in_scenario <- dplyr::anti_join(
+    reference_infilling_df, scenario_df,
     by = c(iso3, ind, year)
   )
 
@@ -182,8 +206,13 @@ recycle_data_scenario_single <- function(df,
     by = c(iso3, ind, year)
   )
 
+  reference_infilling_not_in_covid_shock <- dplyr::anti_join(
+    reference_infilling_not_in_default, covid_shock_df,
+    by = c(iso3, ind, year)
+  )
+
   not_in_scenario <- dplyr::bind_rows(default_not_in_scenario, reported_not_in_default) %>%
-    dplyr::bind_rows(reference_infilling_not_in_default) %>%
+    dplyr::bind_rows(reference_infilling_not_in_covid_shock, covid_shock_not_in_default) %>%
     dplyr::mutate(recycled = TRUE)
 
   if (!include_projection) {
@@ -191,7 +220,7 @@ recycle_data_scenario_single <- function(df,
       dplyr::filter(!.data[[type]] %in% c("imputed", "projected"))
 
     not_in_scenario <- dplyr::bind_rows(not_in_scenario_projs, reported_not_in_default) %>%
-      dplyr::bind_rows(reference_infilling_not_in_default) %>%
+      dplyr::bind_rows(reference_infilling_not_in_covid_shock, covid_shock_not_in_default) %>%
       dplyr::mutate(
         recycled = TRUE,
         !!sym(scenario_col) := scenario
@@ -300,6 +329,7 @@ make_default_scenario <- function(df,
                                   scenario_col = "scenario",
                                   default_scenario = "default",
                                   scenario_reported_estimated = "routine",
+                                  scenario_covid_shock = "covid_shock",
                                   scenario_reference_infilling = "reference_infilling",
                                   include_projection = TRUE,
                                   recycle_campaigns = TRUE,
@@ -307,7 +337,8 @@ make_default_scenario <- function(df,
                                   trim_years = FALSE) {
   assert_columns(df, iso3, ind, value, year, scenario_col, type)
   assert_unique_rows(df, ind, iso3, year, scenario = scenario_col, ind_ids)
-  assert_scenario_in_df(df, c(scenario, scenario_reported_estimated, scenario_reference_infilling), scenario_col)
+
+  base_scenarios <- c(scenario_reported_estimated, scenario_reference_infilling)
 
   if (is.null(billion)) {
     billion <- "all"
@@ -340,6 +371,7 @@ make_default_scenario <- function(df,
       scenario_col = scenario_col,
       default_scenario = default_scenario,
       scenario_reported_estimated = scenario_reported_estimated,
+      scenario_covid_shock = scenario_covid_shock,
       scenario_reference_infilling = scenario_reference_infilling,
       include_projection = include_projection,
       recycle_campaigns = recycle_campaigns,