+++ R/wrangle_data.R
@@ -43,8 +43,8 @@ wrangle_gho_data <- function(df,
         ind
       ),
       "value" := .data[["NumericValue"]],
-      "lower" := .data[["Low"]],
-      "upper" := .data[["High"]],
+      "lower" := as.double(.data[["Low"]]),
+      "upper" := as.double(.data[["High"]]),
       "use_dash" := TRUE,
       "use_calc" := TRUE,
       "source" := ifelse(is.null(source),
@@ -55,14 +55,9 @@ wrangle_gho_data <- function(df,
         NA_character_,
         type
       ),
-      "type_detail" := NA,
-      "other_detail" := .data[["Comments"]],
-      "upload_detail" := NA,
-      "scenario" := ifelse(is.null(scenario),
-                           NA_character_,
-                           scenario
-      ),
-      "scenario_detail" := NA_character_
+      "type_detail" := NA_character_,
+      "other_detail" := as.character(.data[["Comments"]]),
+      "upload_detail" := NA_character_
     ) %>%
     dplyr::filter(whoville::is_who_member(.data[["iso3"]])) %>%
     dplyr::arrange(.data[["iso3"]], .data[["year"]])
@@ -184,30 +179,30 @@ wrangle_gho_rural_urban_data <- function(df,
 
       # If a total value doesn't exist, use the rural/urban indicator name
       "ind" := dplyr::case_when(
-        !is.na(.data[["NumericValue_TOTL"]]) ~ glue::glue("{ind}"),
-        !is.na(.data[["NumericValue_RUR"]]) ~ glue::glue("{ind}_rural"),
-        !is.na(.data[["NumericValue_URB"]]) ~ glue::glue("{ind}_urban")
+        !is.na(.data[["NumericValue_TOTL"]]) ~ .env$ind,
+        !is.na(.data[["NumericValue_RUR"]]) ~ sprintf("%s_rural", .env$ind),
+        !is.na(.data[["NumericValue_URB"]]) ~ sprintf("%s_urban", .env$ind)
       ),
 
       # If a total value doesn't exist, use the rural/urban value
-      "value" := dplyr::case_when(
+      "value" := as.double(dplyr::case_when(
         !!!make_conds(prefixes = c("NumericValue"), suffixes = c("TOTL", "RUR", "URB"))
-      ),
+      )),
 
       # If a total low doesn't exist, use the rural/urban low
-      "lower" := dplyr::case_when(
+      "lower" := as.double(dplyr::case_when(
         !!!make_conds(prefixes = c("Low"), suffixes = c("TOTL", "RUR", "URB"))
-      ),
+      )),
 
       # If a total high doesn't exist, use the rural/urban high
-      "upper" := dplyr::case_when(
+      "upper" := as.double(dplyr::case_when(
         !!!make_conds(prefixes = c("High"), suffixes = c("TOTL", "RUR", "URB"))
-      ),
+      )),
 
       # If a total source doesn't exist, use the rural/urban source
-      "DataSourceDim" := dplyr::case_when(
+      "DataSourceDim" := as.character(dplyr::case_when(
         !!!make_conds(prefixes = c("DataSourceDim"), suffixes = c("TOTL", "RUR", "URB"))
-      ),
+      )),
 
       # If a data source is explicitly provided, override the sources from the DataSourceDim column
       # Follows the same logic as `wrangle_gho_data` by giving priority to explicit source over GHO source
@@ -216,9 +211,9 @@ wrangle_gho_rural_urban_data <- function(df,
       ),
 
       # If a total comment doesn't exist, use the rural/urban comment
-      "other_detail" := dplyr::case_when(
+      "other_detail" := as.character(dplyr::case_when(
         !!!make_conds(prefixes = c("Comments"), suffixes = c("TOTL", "RUR", "URB"))
-      ),
+      )),
 
       # If a type argument is provided, use that; otherwise NA
       "type" := ifelse(
@@ -226,13 +221,8 @@ wrangle_gho_rural_urban_data <- function(df,
       ),
       use_dash = TRUE,
       use_calc = TRUE,
-      type_detail = NA,
-      upload_detail = NA,
-      "scenario" := ifelse(is.null(scenario),
-                           NA_character_,
-                           scenario
-      ),
-      "scenario_detail" := NA_character_
+      type_detail = NA_character_,
+      upload_detail = NA_character_
     ) %>%
     ## Filter out 'mixed' time series
     # Group time series by iso3
@@ -244,7 +234,7 @@ wrangle_gho_rural_urban_data <- function(df,
     # time series
     dplyr::filter(.data[["ind"]] == .data[["ind_mode"]]) %>%
     # Remove unnecessary columns
-    dplyr::select(xmart_cols()) %>%
+    dplyr::select(!c("DataSourceDim", "ind_mode")) %>%
     # Filter to keep only WHO members
     dplyr::filter(whoville::is_who_member(.data[["iso3"]])) %>%
     # Arrange in ascending order of iso3, year
@@ -296,11 +286,11 @@ wrangle_unsd_data <- function(df,
       "iso3" := whoville::codes_to_iso3(.data[["GeoAreaCode"]], type = "m49"),
       "year" := .data[["TimePeriod"]],
       "ind" := .data[["SeriesCode"]],
-      "value" := .data[["Value"]],
-      "lower" := .data[["LowerBound"]],
-      "upper" := .data[["UpperBound"]],
+      "value" := as.double(.data[["Value"]]),
+      "lower" := as.double(.data[["LowerBound"]]),
+      "upper" := as.double(.data[["UpperBound"]]),
       "source" := ifelse(is.null(source),
-        .data[["Source"]],
+        as.character(.data[["Source"]]),
         source
       ),
       "type" := dplyr::case_when(
@@ -308,12 +298,7 @@ wrangle_unsd_data <- function(df,
         .data[["Nature"]] %in% c("C", "CA") ~ "reported",
         .data[["Nature"]] %in% c("E", "M") ~ "estimated"
       ),
-      "other_detail" := .data[["FootNote"]],
-      "scenario" := ifelse(is.null(scenario),
-                           NA_character_,
-                           scenario
-      ),
-      "scenario_detail" := NA_character_
+      "other_detail" := as.character(.data[["FootNote"]])
     ) %>%
     dplyr::filter(whoville::is_who_member(.data[["iso3"]])) %>%
     dplyr::arrange(.data[["iso3"]], .data[["year"]])